@@ -30,8 +30,7 @@
 #define TESSERACT_RVIZ_INTERACTIVE_MARKER_INTERACTIVE_MARKER_CONTROL_H
 
 #ifndef Q_MOC_RUN
-#include <boost/shared_ptr.hpp>
-#include <boost/enable_shared_from_this.hpp>
+#include <memory>
 
 #include <OgreRay.h>
 #include <OgreVector3.h>
@@ -43,22 +42,26 @@
 
 #include <QCursor>
 
-#include "rviz/selection/forwards.h"
-#include "rviz/viewport_mouse_event.h"
-#include "rviz/interactive_object.h"
+#include <rviz_common/interaction/forwards.hpp>
+#include <rviz_common/viewport_mouse_event.hpp>
+#include <rviz_common/interactive_object.hpp>
 
 namespace Ogre
 {
 class SceneNode;
 }
 
-namespace rviz
+namespace rviz_common
 {
-class DisplayContext;
+//class DisplayContext;
+}  // namespace rviz_common
+
+namespace rviz_rendering
+{
 class PointsMarker;
 class Line;
 class Shape;
-}  // namespace rviz
+}
 
 namespace tesseract_rviz
 {
@@ -94,8 +97,8 @@ enum class OrientationMode
  * A single control element of an InteractiveMarker.
  */
 class InteractiveMarkerControl : public Ogre::SceneManager::Listener,
-                                 public rviz::InteractiveObject,
-                                 public boost::enable_shared_from_this<InteractiveMarkerControl>
+                                 public rviz_common::InteractiveObject,
+                                 public std::enable_shared_from_this<InteractiveMarkerControl>
 {
 public:
   using Ptr = boost::shared_ptr<InteractiveMarkerControl>;
@@ -103,7 +106,7 @@ public:
 
   InteractiveMarkerControl(const std::string& name,
                            const std::string& description,
-                           rviz::DisplayContext* context,
+                           rviz_common::DisplayContext* context,
                            Ogre::SceneNode* reference_node,
                            InteractiveMarker* parent,
                            const InteractiveMode interactive_mode,
@@ -123,13 +126,13 @@ public:
    * @brief Add marker to the controller
    * @param marker to add to the controller
    */
-  void addMarker(MarkerBase::Ptr marker);
+  void addMarker(std::shared_ptr<MarkerBase> marker);
 
   // called when interactive mode is globally switched on/off
   virtual void enableInteraction(bool enable);
 
   // will receive all mouse events while the handler has focus
-  virtual void handleMouseEvent(rviz::ViewportMouseEvent& event);
+  virtual void handleMouseEvent(rviz_common::ViewportMouseEvent& event);
 
   /**
    * This is the main entry-point for interaction using a 3D cursor.
@@ -156,7 +159,7 @@ public:
    * @param  cursor_rot   The world-relative orientation of the 3D cursor.
    * @param  control_name The name of the child InteractiveMarkerControl calling this function.
    */
-  virtual void handle3DCursorEvent(rviz::ViewportMouseEvent event,
+  virtual void handle3DCursorEvent(rviz_common::ViewportMouseEvent event,
                                    const Ogre::Vector3& cursor_3D_pos,
                                    const Ogre::Quaternion& cursor_3D_orientation);
 
@@ -222,7 +225,7 @@ public:
    * @brief Get the display context
    * @return Display context
    */
-  rviz::DisplayContext* getDisplayContext() { return context_; }
+  rviz_common::DisplayContext* getDisplayContext() { return context_; }
 
   /**
    * @brief Get the size of the ineractive control
@@ -250,29 +253,29 @@ protected:
 
   /** calculate a mouse ray in the reference frame.
    *  A mouse ray is a ray starting at the camera and pointing towards the mouse position. */
-  Ogre::Ray getMouseRayInReferenceFrame(const rviz::ViewportMouseEvent& event, int x, int y);
+  Ogre::Ray getMouseRayInReferenceFrame(const rviz_common::ViewportMouseEvent& event, int x, int y);
 
   /** begin a relative-motion drag. */
-  void beginRelativeMouseMotion(const rviz::ViewportMouseEvent& event);
+  void beginRelativeMouseMotion(const rviz_common::ViewportMouseEvent& event);
 
   /** get the relative motion of the mouse, and put the mouse back
    *  where it was when beginRelativeMouseMotion() was called. */
-  bool getRelativeMouseMotion(const rviz::ViewportMouseEvent& event, int& dx, int& dy);
+  bool getRelativeMouseMotion(const rviz_common::ViewportMouseEvent& event, int& dx, int& dy);
 
   /** Rotate the pose around the camera-frame XY (right/up) axes, based on relative mouse movement. */
-  void rotateXYRelative(const rviz::ViewportMouseEvent& event);
+  void rotateXYRelative(const rviz_common::ViewportMouseEvent& event);
 
   /** Rotate the pose around the camera-frame Z (look) axis, based on relative mouse movement. */
-  void rotateZRelative(const rviz::ViewportMouseEvent& event);
+  void rotateZRelative(const rviz_common::ViewportMouseEvent& event);
 
   /** Move the pose along the mouse ray, based on relative mouse movement. */
-  void moveZAxisRelative(const rviz::ViewportMouseEvent& event);
+  void moveZAxisRelative(const rviz_common::ViewportMouseEvent& event);
 
   /** Move the pose along the mouse ray, based on mouse wheel movement. */
-  void moveZAxisWheel(const rviz::ViewportMouseEvent& event);
+  void moveZAxisWheel(const rviz_common::ViewportMouseEvent& event);
 
   /** Move the pose around the XY view plane (perpendicular to the camera direction). */
-  void moveViewPlane(Ogre::Ray& mouse_ray, const rviz::ViewportMouseEvent& event);
+  void moveViewPlane(Ogre::Ray& mouse_ray, const rviz_common::ViewportMouseEvent& event);
 
   /** Rotate the pose around the local X-axis, following the mouse movement.
    *  mouse_ray is relative to the reference frame. */
@@ -297,7 +300,7 @@ protected:
 
   /** Translate along the local X-axis, following the mouse movement.
    *  mouse_ray is relative to the reference frame. */
-  void moveAxis(const Ogre::Ray& mouse_ray, const rviz::ViewportMouseEvent& event);
+  void moveAxis(const Ogre::Ray& mouse_ray, const rviz_common::ViewportMouseEvent& event);
 
   /** Translate along the local X-axis, following the 3D cursor movement. */
   void moveAxis(const Ogre::Vector3& cursor_position_in_reference_frame);
@@ -346,16 +349,16 @@ protected:
   // Save a copy of the latest mouse event with the event type set to
   // QEvent::MouseMove, so that update() can resend the mouse event during
   // drag actions to maintain consistent behavior.
-  void recordDraggingInPlaceEvent(rviz::ViewportMouseEvent& event);
+  void recordDraggingInPlaceEvent(rviz_common::ViewportMouseEvent& event);
 
   // Begin a new mouse motion.  Called when left button is pressed to begin a drag.
-  void beginMouseMovement(rviz::ViewportMouseEvent& event, bool line_visible);
+  void beginMouseMovement(rviz_common::ViewportMouseEvent& event, bool line_visible);
 
   // Motion part of mouse event handling.
-  void handleMouseMovement(rviz::ViewportMouseEvent& event);
+  void handleMouseMovement(rviz_common::ViewportMouseEvent& event);
 
   // Mouse wheel part of mouse event handling.
-  void handleMouseWheelMovement(rviz::ViewportMouseEvent& event);
+  void handleMouseWheelMovement(rviz_common::ViewportMouseEvent& event);
 
   // Return closest point on a line to a test point.
   Ogre::Vector3 closestPointOnLineToPoint(const Ogre::Vector3& line_start,
@@ -369,11 +372,11 @@ protected:
   bool mouse_dragging_;
   Ogre::Viewport* drag_viewport_;
 
-  rviz::ViewportMouseEvent dragging_in_place_event_;
+//  rviz_common::ViewportMouseEvent dragging_in_place_event_;  // TODO: replace?
 
-  rviz::DisplayContext* context_;
+  rviz_common::DisplayContext* context_;
 
-  rviz::CollObjectHandle coll_object_handle_;
+//  rviz_common::CollObjectHandle coll_object_handle_;  // TODO replace?
 
   /** Node representing reference frame in tf, like /map, /base_link,
    * /head, etc.  Same as the field in InteractiveMarker. */
@@ -411,7 +414,7 @@ protected:
 
   std::string name_;
 
-  std::vector<boost::shared_ptr<MarkerBase>> markers_;
+  std::vector<std::shared_ptr<MarkerBase>> markers_;
 
   InteractiveMarker* parent_;
 
