@@ -40,7 +40,7 @@ class Vector3;
 class ColourValue;
 }  // namespace Ogre
 
-namespace rviz
+namespace rviz_common
 {
 class DisplayContext;
 }
